only update prev and presence column if write succeeds
This fixes a couple bugs that could have been caused by failed
writes. Previously, if a write failed (however unlikely that is) then
the `prev` field in the encoder would still have been changed, and so
future writes would have been corrupted. The same is true for updating
the presence column - now it only happens after the write succeeds.